beryllium: shims: rename libpiex_shim to avoid conflict
The current name has conflict with the one in hardware/custom/compat.
I chose not to use that shim because mangling rules are not part of
the c++ standard. A shim should always use the exposed symbol ABI
directly instead of relying on the compiler to generate correct
mangled symbols from c++ code.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Change-Id: Idf6f270e0424b14d5f0ebf3a943ee24cd196f1ec